new update
Đã tương tác được với database

+++ server/src/main/java/com/pbl3/musicapplication/controller/ArtistController.java
@@ -1,6 +1,10 @@
 package com.pbl3.musicapplication.controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,22 +25,43 @@ public class ArtistController {
     private ArtistService artistService;
 
     @GetMapping("/{id}")
-    public ArtistModel findById(@PathVariable Integer id) {
-        return artistService.findById(id);
+    public ResponseEntity<ArtistModel> findById(@PathVariable Integer id) {
+        return ResponseEntity.ok(artistService.findById(id));
     }
-
-    @PostMapping() 
-    public Artist create(@RequestBody ArtistModel artistModel) {
-        return artistService.create(artistModel);
+    @GetMapping("/all")
+    public ResponseEntity<List<ArtistModel>> findAll() {
+        return ResponseEntity.ok(artistService.findAll());
+    }
+    @GetMapping("/all/name")
+    public ResponseEntity<List<String>> getArtistNameList() {
+        return ResponseEntity.ok(artistService.getArtistNameList());
+    }
+    
+    @PostMapping(consumes = "application/json;charset=UTF-8", produces = "application/json;charset=UTF-8")
+    public ResponseEntity<ArtistModel> create(@RequestBody ArtistModel artistModel) {
+        Artist artist = artistService.create(artistModel);
+        if (artist == null) {
+            return ResponseEntity.badRequest().body(null);
+        }
+        return ResponseEntity.ok(new ArtistModel(artist));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Integer id) {
+    public ResponseEntity<String> deleteById(@PathVariable Integer id) {
+        if (artistService.findById(id) == null)
+            return new ResponseEntity<>("Not found object", HttpStatus.NO_CONTENT);
         artistService.deleteById(id);
+        return new ResponseEntity<>("Deleted", HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("/{id}")
-    public Artist update(@PathVariable Integer id, @RequestBody ArtistModel artistModel) {
-        return artistService.update(id, artistModel);
+    public ResponseEntity<ArtistModel> update(@PathVariable Integer id, @RequestBody ArtistModel artistModel) {
+        Artist artist = artistService.update(id, artistModel);
+        if (artist == null) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(new ArtistModel(artist), HttpStatus.NO_CONTENT);
     }
+
+    
 }
