+++ server/src/main/java/com/pbl3/musicapplication/controller/SongController.java
@@ -1,9 +1,10 @@
 package com.pbl3.musicapplication.controller;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,8 +14,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.pbl3.musicapplication.model.entity.Artist;
 import com.pbl3.musicapplication.model.entity.Song;
 import com.pbl3.musicapplication.model.model.SongModel;
+import com.pbl3.musicapplication.service.ArtistService;
 import com.pbl3.musicapplication.service.SongService;
 
 @RestController
@@ -22,34 +25,55 @@ import com.pbl3.musicapplication.service.SongService;
 public class SongController {
     @Autowired
     private SongService songService;
+    @Autowired
+    private ArtistService artistService;
 
-    @GetMapping("/localDateTime/currentTime")
-    public String getDateTime() {
-        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm:ss");
-        return (LocalDateTime.now()).format(dateTimeFormatter);
-    }
-    @PostMapping("/localDateTime/newTime")
-    public String creaLocalDateTime(@RequestBody String localDateTimeString) {
-        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm:ss");
-        return LocalDateTime.parse(localDateTimeString, dateTimeFormatter).toString();
-    }
     @GetMapping("/{id}")
-    public SongModel findById(@PathVariable Integer id) {
-        return songService.findById(id);
+    public ResponseEntity<SongModel> findById(@PathVariable Integer id) {
+        return ResponseEntity.ok(songService.findById(id));
+    }
+    @GetMapping("/all")
+    public ResponseEntity<List<SongModel>> findAll() {
+        return ResponseEntity.ok(songService.findAll());
+    }
+    @GetMapping("/all/name")
+    public ResponseEntity<List<String>>  getSongNameList() {
+        return ResponseEntity.ok(songService.getSongNameList());
     }
 
-    @PostMapping
-    public Song create(@RequestBody SongModel songModel) {
-        return songService.create(songModel);
+    @PostMapping("/{artistId}")
+    public ResponseEntity<SongModel> create(@PathVariable Integer artistId, @RequestBody SongModel songModel) {
+        if (artistService.findById(artistId) != null) {
+            Song song = songService.create(songModel);
+            if (song == null) {
+                return ResponseEntity.badRequest().body(null);
+            }
+
+            songService.updateArtist(artistId, song.getSongId(), true);
+            return ResponseEntity.ok(new SongModel(song));
+        }
+        else return ResponseEntity.badRequest().body(null);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteById(@PathVariable Integer id) {
+    public ResponseEntity<String> deleteById(@PathVariable Integer id) {
+        if (songService.findById(id) == null)
+            return new ResponseEntity<>("Not found object", HttpStatus.NO_CONTENT);
+        
+        Artist artist = artistService.findSingleAndEpSong(id);
+        if (artist != null) {
+            songService.updateArtist(artist.getArtistId(), id, false);
+        }
         songService.deleteById(id);
+        return new ResponseEntity<>("Deleted", HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("/{id}")
-    public Song update(@PathVariable Integer id, @RequestBody SongModel songModel) {
-        return songService.update(id, songModel);
+    public ResponseEntity<SongModel> update(@PathVariable Integer id, @RequestBody SongModel songModel) {
+        Song song = songService.update(id, songModel);
+        if (song == null) {
+            return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(new SongModel(song), HttpStatus.NO_CONTENT);
     }
 }
